requirements.txt
langchain
google-generativeai
python-dotenv
PyPDF2
qdrant-client